+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix sandy farmland not having a loottable
 - Remove Vanilla Igloos from Shattered Glacier
 - Fix a bunch of Leaves drops
+- Remove Extra ModPlatform Classes, Add getConfigPath to PlatformHandler
 
 # 1.0.2-Beta
 - Fix Golden Apple from Green Apple Recipe overriding Vanilla Recipe